test: tighten failure detector simulation harness
