+++ crates/node/src/failure_detector/fd_state.rs
@@ -697,653 +697,5 @@ impl FdState {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::collections::VecDeque;
-    use std::time::Duration;
-
-    use itertools::Itertools;
-    use restate_types::NodeId;
-    use restate_types::cluster_state::{ClusterState, NodeState as PublishedNodeState};
-    use restate_types::net::address::AdvertisedAddress;
-    use restate_types::nodes_config::{NodeConfig, NodesConfiguration, Role};
-    use restate_types::{GenerationalNodeId, RestateVersion};
-
-    use super::*;
-
-    const A: GenerationalNodeId = GenerationalNodeId::new(1, 1);
-    const B: GenerationalNodeId = GenerationalNodeId::new(2, 1);
-    const C: GenerationalNodeId = GenerationalNodeId::new(3, 1);
-
-    /// A compact, deterministic record of one production-state-machine event.
-    ///
-    /// The test assertions deliberately include the rendered trace, so a failing schedule is
-    /// directly replayable from the event sequence without relying on timing-sensitive logs.
-    #[derive(Debug)]
-    struct TraceEntry {
-        elapsed: Duration,
-        event: String,
-        actor: GenerationalNodeId,
-        full_intervals: Option<u32>,
-        stable: bool,
-        last_direct_receive: Duration,
-        peers: Vec<PeerTrace>,
-    }
-
-    #[derive(Debug)]
-    struct PeerTrace {
-        node_id: GenerationalNodeId,
-        gossip_age: u32,
-        state: NodeState,
-        published_state: PublishedNodeState,
-    }
-
-    #[derive(Debug)]
-    struct FdTrace {
-        started_at: Instant,
-        entries: Vec<TraceEntry>,
-    }
-
-    impl FdTrace {
-        fn new() -> Self {
-            Self {
-                started_at: Instant::now(),
-                entries: Vec::new(),
-            }
-        }
-
-        fn record(
-            &mut self,
-            event: impl Into<String>,
-            actor: GenerationalNodeId,
-            full_intervals: Option<u32>,
-            stable: bool,
-            last_direct_receive: Duration,
-            peers: Vec<PeerTrace>,
-        ) {
-            self.entries.push(TraceEntry {
-                elapsed: self.started_at.elapsed(),
-                event: event.into(),
-                actor,
-                full_intervals,
-                stable,
-                last_direct_receive,
-                peers,
-            });
-        }
-
-        fn render(&self) -> String {
-            self.entries
-                .iter()
-                .enumerate()
-                .map(|(index, entry)| {
-                    let peers = entry
-                        .peers
-                        .iter()
-                        .map(|peer| {
-                            format!(
-                                "{} age={} local={} published={:?}",
-                                peer.node_id,
-                                peer.gossip_age,
-                                state_name(peer.state),
-                                peer.published_state
-                            )
-                        })
-                        .join(", ");
-                    format!(
-                        "#{index} +{:?} actor={} event={} full_intervals={:?} stable={} last_direct_receive={:?} [{}]",
-                        entry.elapsed,
-                        entry.actor,
-                        entry.event,
-                        entry.full_intervals,
-                        entry.stable,
-                        entry.last_direct_receive,
-                        peers
-                    )
-                })
-                .join("\n")
-        }
-    }
-
-    fn state_name(state: NodeState) -> &'static str {
-        match state {
-            NodeState::Alive => "Alive",
-            NodeState::Suspect { .. } => "Suspect",
-            NodeState::Dead => "Dead",
-            NodeState::FailingOver => "FailingOver",
-        }
-    }
-
-    struct FdActor {
-        state: FdState,
-        cluster_state: ClusterState,
-    }
-
-    impl FdActor {
-        fn new(node_id: GenerationalNodeId, nodes_config: &NodesConfiguration) -> Self {
-            let cluster_state = ClusterState::default();
-            let state = FdState::new(
-                node_id,
-                nodes_config,
-                PartitionReplicaSetStates::default(),
-                cluster_state.clone().updater(),
-            );
-            Self {
-                state,
-                cluster_state,
-            }
-        }
-
-        fn peer_state(&self, peer: GenerationalNodeId) -> NodeState {
-            self.state.node_states[&peer.as_plain()].state
-        }
-
-        fn peer_age(&self, peer: GenerationalNodeId) -> u32 {
-            self.state.node_states[&peer.as_plain()].gossip_age
-        }
-
-        fn establish_stable_alive_view(&mut self, opts: &GossipOptions) {
-            self.state.num_gossip_received = opts.gossip_fd_stability_threshold.get() as usize;
-            for node in self.state.node_states.values_mut() {
-                node.gossip_age = 0;
-                node.state = NodeState::Alive;
-                self.state
-                    .cs_updater
-                    .upsert_node_state(node.gen_node_id, node.state.into());
-            }
-        }
-    }
-
-    #[derive(Clone, Copy, Debug)]
-    enum Delivery {
-        Queue,
-        Drop,
-        Duplicate,
-    }
-
-    #[derive(Debug)]
-    struct QueuedGossip {
-        from: GenerationalNodeId,
-        message: Gossip,
-    }
-
-    /// Deterministically drives production `FdState` instances without a Tokio select loop.
-    ///
-    /// A caller explicitly chooses every tick and every network outcome. A paused actor simply
-    /// receives no `tick()` calls while its mailbox continues accumulating gossip.
-    struct FdSimulation {
-        opts: GossipOptions,
-        nodes_config: NodesConfiguration,
-        actors: Vec<FdActor>,
-        mailboxes: Vec<VecDeque<QueuedGossip>>,
-        trace: FdTrace,
-    }
-
-    impl FdSimulation {
-        fn stable_three_node_cluster() -> Self {
-            let opts = GossipOptions::default();
-            let nodes_config = nodes_config();
-            let mut actors = [A, B, C]
-                .into_iter()
-                .map(|node_id| FdActor::new(node_id, &nodes_config))
-                .collect_vec();
-            for actor in &mut actors {
-                actor.establish_stable_alive_view(&opts);
-            }
-
-            Self {
-                opts,
-                nodes_config,
-                actors,
-                mailboxes: (0..3).map(|_| VecDeque::new()).collect(),
-                trace: FdTrace::new(),
-            }
-        }
-
-        fn actor_index(node_id: GenerationalNodeId) -> usize {
-            match node_id {
-                A => 0,
-                B => 1,
-                C => 2,
-                _ => panic!("unexpected test node {node_id}"),
-            }
-        }
-
-        fn actor(&self, node_id: GenerationalNodeId) -> &FdActor {
-            &self.actors[Self::actor_index(node_id)]
-        }
-
-        fn actor_mut(&mut self, node_id: GenerationalNodeId) -> &mut FdActor {
-            &mut self.actors[Self::actor_index(node_id)]
-        }
-
-        fn record(
-            &mut self,
-            event: impl Into<String>,
-            actor: GenerationalNodeId,
-            full_intervals: Option<u32>,
-        ) {
-            let fd_actor = self.actor(actor);
-            let stable = fd_actor.state.is_stable(&self.opts);
-            let last_direct_receive = fd_actor.state.last_gossip_received_at.elapsed();
-            let mut peers = fd_actor
-                .state
-                .node_states
-                .values()
-                .map(|node| PeerTrace {
-                    node_id: node.gen_node_id,
-                    gossip_age: node.gossip_age,
-                    state: node.state,
-                    published_state: fd_actor
-                        .cluster_state
-                        .get_node_state(NodeId::Generational(node.gen_node_id)),
-                })
-                .collect_vec();
-            peers.sort_unstable_by_key(|peer| peer.node_id);
-            self.trace.record(
-                event,
-                actor,
-                full_intervals,
-                stable,
-                last_direct_receive,
-                peers,
-            );
-        }
-
-        async fn advance_one_interval(&mut self) {
-            tokio::time::advance(*self.opts.gossip_tick_interval).await;
-        }
-
-        fn tick(&mut self, node_id: GenerationalNodeId) -> u32 {
-            let opts = self.opts.clone();
-            let full_intervals = self
-                .actor(node_id)
-                .state
-                .last_gossip_tick
-                .elapsed()
-                .div_duration_f32(*opts.gossip_tick_interval)
-                .floor() as u32;
-            let interval_passed = {
-                let actor = self.actor_mut(node_id);
-                let interval_passed = actor.state.gossip_tick(&opts);
-                if actor.state.is_stable(&opts) {
-                    actor.state.detect_peer_failures(&opts);
-                }
-                interval_passed
-            };
-            self.record(
-                format!("tick(interval_passed={interval_passed})"),
-                node_id,
-                Some(full_intervals),
-            );
-            full_intervals
-        }
-
-        fn send_gossip(
-            &mut self,
-            from: GenerationalNodeId,
-            to: GenerationalNodeId,
-            delivery: Delivery,
-        ) {
-            let opts = self.opts.clone();
-            let message = self.actors[Self::actor_index(from)]
-                .state
-                .make_gossip_message(&opts, false, &self.nodes_config);
-            let target = Self::actor_index(to);
-            match delivery {
-                Delivery::Queue => {
-                    self.mailboxes[target].push_back(QueuedGossip { from, message });
-                }
-                Delivery::Drop => {}
-                Delivery::Duplicate => {
-                    self.mailboxes[target].push_back(QueuedGossip {
-                        from,
-                        message: message.clone(),
-                    });
-                    self.mailboxes[target].push_back(QueuedGossip { from, message });
-                }
-            }
-            self.record(format!("gossip {from} -> {to} {delivery:?}"), to, None);
-        }
-
-        fn deliver_next_from(&mut self, to: GenerationalNodeId, from: GenerationalNodeId) {
-            let opts = self.opts.clone();
-            let nodes_config_version = self.nodes_config.version();
-            let target = Self::actor_index(to);
-            let position = self.mailboxes[target]
-                .iter()
-                .position(|queued| queued.from == from)
-                .expect("the requested gossip message is queued");
-            let queued = self.mailboxes[target]
-                .remove(position)
-                .expect("mailbox entry exists");
-            {
-                let actor = self.actor_mut(to);
-                assert!(actor.state.can_admit_message(
-                    &opts,
-                    queued.from,
-                    nodes_config_version,
-                    &queued.message,
-                ));
-                actor.state.update_from_gossip_message(
-                    &opts,
-                    queued.from,
-                    nodes_config_version,
-                    queued.message,
-                );
-            }
-            self.record(format!("deliver gossip {from} -> {to}"), to, None);
-        }
-
-        fn trace(&self) -> String {
-            self.trace.render()
-        }
-    }
-
-    fn nodes_config() -> NodesConfiguration {
-        let mut nodes_config = NodesConfiguration::new_for_testing();
-        for node_id in [A, B, C] {
-            nodes_config.upsert_node(
-                NodeConfig::builder()
-                    .name(format!("node-{node_id}"))
-                    .current_generation(node_id)
-                    .address(AdvertisedAddress::default())
-                    .roles(Role::Admin | Role::Worker)
-                    .binary_version(RestateVersion::current())
-                    .build(),
-            );
-        }
-        nodes_config
-    }
-
-    async fn pause_a_while_b_and_c_continue(sim: &mut FdSimulation, intervals: u32) {
-        for _ in 0..intervals {
-            sim.advance_one_interval().await;
-            keep_b_and_c_healthy(sim);
-            sim.send_gossip(B, A, Delivery::Queue);
-            sim.send_gossip(C, A, Delivery::Queue);
-        }
-    }
-
-    fn keep_b_and_c_healthy(sim: &mut FdSimulation) {
-        sim.tick(B);
-        sim.tick(C);
-        sim.send_gossip(B, C, Delivery::Queue);
-        sim.deliver_next_from(C, B);
-        sim.send_gossip(C, B, Delivery::Queue);
-        sim.deliver_next_from(B, C);
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn paused_observer_bulk_ages_healthy_peers_to_dead() {
-        let mut sim = FdSimulation::stable_three_node_cluster();
-
-        pause_a_while_b_and_c_continue(&mut sim, 12).await;
-
-        assert_eq!(sim.tick(A), 12, "trace:\n{}", sim.trace());
-        assert_eq!(sim.actor(A).peer_age(B), 12, "trace:\n{}", sim.trace());
-        assert_eq!(sim.actor(A).peer_age(C), 12, "trace:\n{}", sim.trace());
-        assert_eq!(
-            sim.actor(A).peer_state(B),
-            NodeState::Dead,
-            "trace:\n{}",
-            sim.trace()
-        );
-        assert_eq!(
-            sim.actor(A).peer_state(C),
-            NodeState::Dead,
-            "trace:\n{}",
-            sim.trace()
-        );
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn every_failure_significant_pause_in_the_incident_range_false_kills_healthy_peers() {
-        for intervals in 12..=30 {
-            let mut sim = FdSimulation::stable_three_node_cluster();
-            pause_a_while_b_and_c_continue(&mut sim, intervals).await;
-
-            assert_eq!(sim.tick(A), intervals, "trace:\n{}", sim.trace());
-            assert_eq!(
-                sim.actor(A).peer_state(B),
-                NodeState::Dead,
-                "pause={intervals}; trace:\n{}",
-                sim.trace()
-            );
-            assert_eq!(
-                sim.actor(A).peer_state(C),
-                NodeState::Dead,
-                "pause={intervals}; trace:\n{}",
-                sim.trace()
-            );
-        }
-    }
-
-    #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-    enum ResumeEvent {
-        DeliverB,
-        DeliverC,
-        DelayedTick,
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn resume_order_permutations_produce_divergent_views() {
-        for (schedule, expected) in [
-            (
-                [
-                    ResumeEvent::DeliverB,
-                    ResumeEvent::DeliverC,
-                    ResumeEvent::DelayedTick,
-                ],
-                ("Dead", "Dead"),
-            ),
-            (
-                [
-                    ResumeEvent::DeliverB,
-                    ResumeEvent::DelayedTick,
-                    ResumeEvent::DeliverC,
-                ],
-                ("Suspect", "Suspect"),
-            ),
-            (
-                [
-                    ResumeEvent::DeliverC,
-                    ResumeEvent::DeliverB,
-                    ResumeEvent::DelayedTick,
-                ],
-                ("Dead", "Dead"),
-            ),
-            (
-                [
-                    ResumeEvent::DeliverC,
-                    ResumeEvent::DelayedTick,
-                    ResumeEvent::DeliverB,
-                ],
-                ("Suspect", "Suspect"),
-            ),
-            (
-                [
-                    ResumeEvent::DelayedTick,
-                    ResumeEvent::DeliverB,
-                    ResumeEvent::DeliverC,
-                ],
-                ("Suspect", "Suspect"),
-            ),
-            (
-                [
-                    ResumeEvent::DelayedTick,
-                    ResumeEvent::DeliverC,
-                    ResumeEvent::DeliverB,
-                ],
-                ("Suspect", "Suspect"),
-            ),
-        ] {
-            let mut sim = FdSimulation::stable_three_node_cluster();
-            pause_a_while_b_and_c_continue(&mut sim, 12).await;
-
-            for event in schedule {
-                match event {
-                    ResumeEvent::DeliverB => sim.deliver_next_from(A, B),
-                    ResumeEvent::DeliverC => sim.deliver_next_from(A, C),
-                    ResumeEvent::DelayedTick => {
-                        assert_eq!(sim.tick(A), 12, "trace:\n{}", sim.trace())
-                    }
-                }
-            }
-
-            sim.advance_one_interval().await;
-            assert_eq!(sim.tick(A), 1, "trace:\n{}", sim.trace());
-            assert_eq!(
-                (
-                    state_name(sim.actor(A).peer_state(B)),
-                    state_name(sim.actor(A).peer_state(C)),
-                ),
-                expected,
-                "schedule={schedule:?}; trace:\n{}",
-                sim.trace()
-            );
-        }
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn transport_loss_ages_a_scheduled_observer_one_interval_at_a_time() {
-        let mut sim = FdSimulation::stable_three_node_cluster();
-
-        for expected_age in 1..=11 {
-            sim.advance_one_interval().await;
-            assert_eq!(sim.tick(A), 1, "trace:\n{}", sim.trace());
-            assert_eq!(
-                sim.actor(A).peer_age(B),
-                expected_age,
-                "trace:\n{}",
-                sim.trace()
-            );
-            assert_eq!(
-                sim.actor(A).peer_state(B),
-                if expected_age == 11 {
-                    NodeState::Dead
-                } else {
-                    NodeState::Alive
-                },
-                "trace:\n{}",
-                sim.trace()
-            );
-        }
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn paused_observer_cannot_distinguish_a_silent_peer_from_a_healthy_queued_peer() {
-        let mut sim = FdSimulation::stable_three_node_cluster();
-
-        for _ in 0..12 {
-            sim.advance_one_interval().await;
-            sim.tick(C);
-            sim.send_gossip(C, A, Delivery::Queue);
-        }
-
-        assert_eq!(sim.tick(A), 12, "trace:\n{}", sim.trace());
-        assert_eq!(
-            sim.actor(A).peer_state(B),
-            NodeState::Dead,
-            "trace:\n{}",
-            sim.trace()
-        );
-        assert_eq!(
-            sim.actor(A).peer_state(C),
-            NodeState::Dead,
-            "trace:\n{}",
-            sim.trace()
-        );
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn delayed_ticks_create_a_dead_suspect_dead_flap_that_eventually_converges() {
-        let mut sim = FdSimulation::stable_three_node_cluster();
-
-        pause_a_while_b_and_c_continue(&mut sim, 12).await;
-        sim.tick(A);
-        assert_eq!(
-            sim.actor(A).peer_state(B),
-            NodeState::Dead,
-            "trace:\n{}",
-            sim.trace()
-        );
-
-        sim.deliver_next_from(A, B);
-        sim.advance_one_interval().await;
-        sim.tick(A);
-        assert!(
-            matches!(sim.actor(A).peer_state(B), NodeState::Suspect { .. }),
-            "trace:\n{}",
-            sim.trace()
-        );
-
-        for _ in 0..12 {
-            sim.advance_one_interval().await;
-        }
-        sim.tick(A);
-        assert_eq!(
-            sim.actor(A).peer_state(B),
-            NodeState::Dead,
-            "trace:\n{}",
-            sim.trace()
-        );
-
-        sim.send_gossip(B, A, Delivery::Queue);
-        sim.deliver_next_from(A, B);
-        sim.advance_one_interval().await;
-        sim.tick(A);
-        assert!(matches!(
-            sim.actor(A).peer_state(B),
-            NodeState::Suspect { .. }
-        ));
-
-        for _ in 0..50 {
-            sim.advance_one_interval().await;
-            keep_b_and_c_healthy(&mut sim);
-            sim.send_gossip(B, A, Delivery::Queue);
-            sim.deliver_next_from(A, B);
-            sim.tick(A);
-        }
-        assert_eq!(
-            sim.actor(A).peer_state(B),
-            NodeState::Alive,
-            "trace:\n{}",
-            sim.trace()
-        );
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn duplicate_gossip_is_explicitly_replayable_without_changing_a_live_view() {
-        let mut sim = FdSimulation::stable_three_node_cluster();
-
-        sim.send_gossip(B, A, Delivery::Duplicate);
-        sim.deliver_next_from(A, B);
-        sim.deliver_next_from(A, B);
-
-        assert_eq!(
-            sim.actor(A).peer_state(B),
-            NodeState::Alive,
-            "trace:\n{}",
-            sim.trace()
-        );
-        assert_eq!(sim.actor(A).peer_age(B), 0, "trace:\n{}", sim.trace());
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn dropped_gossip_is_explicitly_replayable() {
-        let mut sim = FdSimulation::stable_three_node_cluster();
-
-        sim.send_gossip(B, A, Delivery::Drop);
-        for _ in 0..11 {
-            sim.advance_one_interval().await;
-            sim.tick(A);
-        }
-
-        assert_eq!(
-            sim.actor(A).peer_state(B),
-            NodeState::Dead,
-            "trace:\n{}",
-            sim.trace()
-        );
-    }
-}
+#[path = "fd_state/tests.rs"]
+mod tests;
