+++ crates/node/src/failure_detector/node_state.rs
@@ -106,6 +106,8 @@ pub struct Node {
     // The connection swims on gossip's swimlane.
     #[debug("is_closed?={}, is_none?={}", connection.as_ref().is_some_and(|c| c.is_closed()), connection.is_none())]
     connection: Option<LazyConnection>,
+    #[cfg(test)]
+    terminally_closed_for_test: bool,
 }
 
 impl Node {
@@ -118,6 +120,8 @@ impl Node {
             in_failover: false,
             nc_version_witness: Version::INVALID,
             connection: None,
+            #[cfg(test)]
+            terminally_closed_for_test: false,
         }
     }
     /// Resets the node state if the generation is higher than the current one.
@@ -138,6 +142,10 @@ impl Node {
             self.in_failover = false;
             self.nc_version_witness = nc_version;
             self.connection = None;
+            #[cfg(test)]
+            {
+                self.terminally_closed_for_test = false;
+            }
             true
         } else if instance_ts > self.instance_ts {
             // note that instance_ts can be 0 if the node is not alive yet, we accept that we don't
@@ -286,7 +294,21 @@ impl Node {
     }
 
     pub fn is_gone(&self) -> bool {
-        self.connection.as_ref().is_some_and(|c| c.is_closed())
+        self.connection.as_ref().is_some_and(|c| c.is_closed()) || {
+            #[cfg(test)]
+            {
+                self.terminally_closed_for_test
+            }
+            #[cfg(not(test))]
+            {
+                false
+            }
+        }
+    }
+
+    #[cfg(test)]
+    pub(super) fn mark_terminally_closed_for_test(&mut self) {
+        self.terminally_closed_for_test = true;
     }
 
     pub fn send_gossip(
